refactor: extract KitGrid components to achieve <300 LOC target
